LPS-84964 Applies test pattern to dynamic-data-mapping-form-field-type
