+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/test/metal/js/FieldBase/FieldBase.es.js
@@ -1,4 +1,4 @@
-import FieldBase from '../FieldBase.es';
+import FieldBase from 'source/FieldBase/FieldBase.es';
 
 let component;
 const spritemap = 'icons.svg';
