+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/test/metal/js/Grid/Grid.es.js
@@ -1,10 +1,10 @@
-import Date from '../Date.es';
+import Grid from 'source/Grid/Grid.es';
 
 let component;
 const spritemap = 'icons.svg';
 
 describe(
-	'Field Date',
+	'Grid',
 	() => {
 		afterEach(
 			() => {
@@ -15,9 +15,46 @@ describe(
 		);
 
 		it(
+			'should render columns',
+			() => {
+				component = new Grid(
+					{
+						columns: [
+							{
+								label: 'col1',
+								value: 'fieldId'
+							},
+							{
+								label: 'col2',
+								value: 'fieldId'
+							}
+						],
+						spritemap
+					}
+				);
+
+				expect(component).toMatchSnapshot();
+			}
+		);
+
+		it(
+			'should render no columns when columns comes empty',
+			() => {
+				component = new Grid(
+					{
+						columns: [],
+						spritemap
+					}
+				);
+
+				expect(component).toMatchSnapshot();
+			}
+		);
+
+		it(
 			'should be not edidable',
 			() => {
-				component = new Date(
+				component = new Grid(
 					{
 						readOnly: false,
 						spritemap
@@ -31,7 +68,7 @@ describe(
 		it(
 			'should have a tip',
 			() => {
-				component = new Date(
+				component = new Grid(
 					{
 						spritemap,
 						tip: 'Type something'
@@ -45,7 +82,7 @@ describe(
 		it(
 			'should have an id',
 			() => {
-				component = new Date(
+				component = new Grid(
 					{
 						id: 'ID',
 						spritemap
@@ -59,7 +96,7 @@ describe(
 		it(
 			'should have a label',
 			() => {
-				component = new Date(
+				component = new Grid(
 					{
 						label: 'label',
 						spritemap
@@ -71,23 +108,9 @@ describe(
 		);
 
 		it(
-			'should have a placeholder',
-			() => {
-				component = new Date(
-					{
-						placeholder: '__/__/____',
-						spritemap
-					}
-				);
-
-				expect(component).toMatchSnapshot();
-			}
-		);
-
-		it(
 			'should not be required',
 			() => {
-				component = new Date(
+				component = new Grid(
 					{
 						required: false,
 						spritemap
@@ -99,9 +122,46 @@ describe(
 		);
 
 		it(
+			'should render rows',
+			() => {
+				component = new Grid(
+					{
+						rows: [
+							{
+								label: 'row1',
+								value: 'fieldId'
+							},
+							{
+								label: 'row2',
+								value: 'fieldId'
+							}
+						],
+						spritemap
+					}
+				);
+
+				expect(component).toMatchSnapshot();
+			}
+		);
+
+		it(
+			'should render no rows when row comes empty',
+			() => {
+				component = new Grid(
+					{
+						rows: [],
+						spritemap
+					}
+				);
+
+				expect(component).toMatchSnapshot();
+			}
+		);
+
+		it(
 			'should render Label if showLabel is true',
 			() => {
-				component = new Date(
+				component = new Grid(
 					{
 						label: 'text',
 						showLabel: true,
@@ -116,23 +176,9 @@ describe(
 		it(
 			'should have a spritemap',
 			() => {
-				component = new Date(
+				component = new Grid(
 					{
 						spritemap
-					}
-				);
-
-				expect(component).toMatchSnapshot();
-			}
-		);
-
-		it(
-			'should have a value',
-			() => {
-				component = new Date(
-					{
-						spritemap,
-						value: 'value'
 					}
 				);
